pyproject.toml
[project]
name = "software-prac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "aio-pika>=9.5.5",
    "cassandra-driver>=3.29.2",
    "msgpack>=1.1.0",
]
